refactor: 타입 변경

+++ src/main/java/com/example/urikkiriserver/domain/user/presentation/dto/response/UserRankingResponseElement.java
@@ -3,9 +3,9 @@ package com.example.urikkiriserver.domain.user.presentation.dto.response;
 public record UserRankingResponseElement(
         Integer level,
         String nickname,
-        Long bananaxp
+        Integer bananaxp
 ) {
-    public static UserRankingResponseElement of(Integer level, String nickname, Long bananaxp) {
+    public static UserRankingResponseElement of(Integer level, String nickname, Integer bananaxp) {
         return new UserRankingResponseElement(level, nickname, bananaxp);
     }
 }
